+++ src/services/repository/OrderRepository.php
@@ -14,7 +14,7 @@ use Craft;
 use Exception;
 use craft\db\Query;
 use craft\helpers\Db;
-
+use craft\records\Element;
 use craft\elements\db\ElementQuery;
 use acclaro\translations\elements\Order;
 use acclaro\translations\records\OrderRecord;
@@ -29,6 +29,15 @@ class OrderRepository
     public function getOrderById($orderId)
     {
         return Craft::$app->elements->getElementById($orderId);
+    }
+
+    /**
+     * @param  int|string $orderId
+     * @return \acclaro\translations\elements\Order|null
+     */
+    public function getOrderByIdWithTrashed($orderId)
+    {
+        return Element::findOne(['id' => $orderId]);
     }
 
     /**